refactored scoring, final score working better
